backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn==0.24.0
alembic==1.13.1
openai==1.3.5
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
asyncpg==0.29.0